requirements_verl.txt
absl-py==2.3.1
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
agentscope==1.0.6
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anthropic==0.62.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
beast-logger==0.0.15
blake3==1.0.5
blobfile==3.0.0
boto3==1.40.6
botocore==1.40.6
build==1.3.0
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.8.3
cffi==2.0.0b1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
compressed-tensors==0.10.2
cuda-bindings==13.0.0
cuda-pathfinder==1.1.0
cuda-python==13.0.0
cupy-cuda12x==13.5.1
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
deprecated==1.2.18
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.18.0
flashinfer-python==0.2.9rc2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0rc0
hydra-core==1.3.2
identify==2.6.13
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
interegular==0.3.3
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lark==1.2.2
latex2sympy2-extended==1.10.2
liger-kernel==0.6.1
litellm==1.75.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
lxml==6.0.0
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
math-verify==0.8.0
mathruler==0.1.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.3
modelscope==1.28.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-frontend==1.13.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.90.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.12
orjson==3.11.1
outlines==0.1.11
outlines-core==0.2.10
packaging==25.0
pandas==2.3.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.17.0
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==4.3.0
prettytable==3.16.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pybind11==3.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.12.0a1
pydantic-core==2.37.2
pydantic-extra-types==2.10.5
pyecharts==2.0.8
pyext==0.7
pygments==2.19.2
pylatexenc==2.10
pynvml==12.0.0
pyproject-hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
pyyaml==6.0.2
pyzmq==27.0.1
qwen-vl-utils==0.0.11
ray==2.48.0
referencing==0.36.2
regex==2025.7.33
requests==2.32.4
rich==13.9.4
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.8
s3transfer==0.13.1
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
setproctitle==1.3.6
setuptools==80.9.0
sgl-kernel==0.2.4
sglang==0.4.9.post6
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.47.2
swankit==0.2.4
swanlab==0.6.8
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.9.1
tiktoken==0.11.0
timm==1.0.16
tokenizers==0.21.4
tomli==2.2.1
torch==2.7.1
torch-memory-saver==0.0.8
torchao==0.9.0
torchaudio==2.7.1
torchdata==0.11.0
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
triton==3.3.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.33.1
vllm==0.10.0
wandb==0.21.1
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
debugpy>=1.8.0